refactor(): VerifierGame became an Use Case - clean code
