+++ app/src/main/java/com/rossel/android/sdk/mytictactoe/domain/interfaces/IState.kt
@@ -1,3 +1,3 @@
 package com.rossel.android.sdk.mytictactoe.domain.interfaces
 
-interface IState : IStateMove, IStateBoard, IStateVerifier
+interface IState : IStateMove, IStateBoard
