--- app/src/main/java/com/rossel/android/sdk/mytictactoe/domain/interfaces/IStateVerifier.kt
@@ -1,6 +1,0 @@
-package com.rossel.android.sdk.mytictactoe.domain.interfaces
-
-
-interface IStateVerifier {
-    fun verifier(): IVerifier
-}
