+++ app/src/main/java/com/rossel/android/sdk/mytictactoe/domain/interfaces/IVerifierUseCase.kt
@@ -3,6 +3,6 @@ package com.rossel.android.sdk.mytictactoe.domain.interfaces
 import com.rossel.android.sdk.mytictactoe.domain.enums.Player
 import com.rossel.android.sdk.mytictactoe.domain.enums.StateEnum
 
-interface IVerifier {
+interface IVerifierUseCase {
     fun verify(board: List<Player>): StateEnum
 }
