+++ app/src/main/java/com/rossel/android/sdk/mytictactoe/domain/usecase/VerifierGameUseCase.kt
@@ -1,12 +1,15 @@
-package com.rossel.android.sdk.mytictactoe.domain.entity
+package com.rossel.android.sdk.mytictactoe.domain.usecase
 
+import com.rossel.android.sdk.mytictactoe.domain.entity.VerifierAntidiagonalRow
+import com.rossel.android.sdk.mytictactoe.domain.entity.VerifierDiagonalRow
+import com.rossel.android.sdk.mytictactoe.domain.entity.VerifierHorizontalRow
+import com.rossel.android.sdk.mytictactoe.domain.entity.VerifierVerticalRow
 import com.rossel.android.sdk.mytictactoe.domain.enums.Player
 import com.rossel.android.sdk.mytictactoe.domain.enums.StateEnum
-import com.rossel.android.sdk.mytictactoe.domain.interfaces.IVerifier
+import com.rossel.android.sdk.mytictactoe.domain.interfaces.IVerifierUseCase
 import com.rossel.android.sdk.mytictactoe.domain.interfaces.IVerifierRow
 
-
-class VerifierGame: IVerifier {
+class VerifierGameUseCase: IVerifierUseCase {
     private val verifierHorizontalRow: IVerifierRow = VerifierHorizontalRow()
     private val verifierVerticalRow: IVerifierRow = VerifierVerticalRow()
     private val verifierAntidiagonalRow: IVerifierRow = VerifierAntidiagonalRow()
